Rework the idx babel transform to be more reliable
Now, instead of searching for calls to the `idx` function, it searches
for `require('idx')` or `import idx from 'idx'` calls, finds all
references to the assigned variable and transform them.

This also allows us to remove the require() call, which is no longer
needed (and in fact is causing issues when using `haste` in RN due to not
being able to find the `idx` module (since it no longer has the
`@providesModule` annotation).