Enhance TypeScript types
This will allow TypeScript with strictNullCheck flag on to work with
this package. We're making the `prop` argument in the accessor function
a deeply required object so we can safely access it deeply.

I've also added a test file for TypeScript definition so we can make
sure changes to the definition is not breaking them. Currently tests are
not run with any sort of CI. If adding TypeScript as dependency is
desired we can add thtat to test scripts as well